src: library rearrange make lib not have actual functions, and add integration_tests
