+++ .github/workflows/test.yaml
@@ -105,17 +105,18 @@ jobs:
       - name: example 1
         run: |
           ${{ env.RUN_CMD }} run  --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --output-dir output/example1 --populations "AY.4.2*,BA.5.2,XBC.1.6*,XBB.1.5.1,XBL"
-          ${{ env.RUN_CMD }} plot --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --output-dir output/example1
+          ${{ env.RUN_CMD }} plot --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --run-dir output/example1
 
       - name: example 2
         run: |
           ${{ env.RUN_CMD }} run  --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --output-dir output/example2 --alignment data/example2.fasta
-          ${{ env.RUN_CMD }} plot --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --output-dir output/example2
+          ${{ env.RUN_CMD }} plot --dataset-dir dataset/sars-cov-2/${{ env.DATASET_TAG }} --run-dir output/example2
 
       - name: toy1
         run: |
+          ${{ env.RUN_CMD }} dataset download --output-dir dataset/toy1 --name toy1 --tag custom
           ${{ env.RUN_CMD }} run  --dataset-dir dataset/toy1 --output-dir output/toy1 --populations "*" --mask 0,0 --min-length 1
-          ${{ env.RUN_CMD }} plot --dataset-dir dataset/toy1 --output-dir output/toy1
+          ${{ env.RUN_CMD }} plot --dataset-dir dataset/toy1 --run-dir output/toy1
 
       - name: upload output
         uses: actions/upload-artifact@v3
