+++ Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "rebar"
-version = "0.1.0"
+version = "0.2.0"
 edition = "2021"
 
 [dependencies]
@@ -15,7 +15,7 @@ indicatif       = { version = "0.17.5",  default-features = false, features = ["
 indoc           = { version = "2.0.4",   default-features = false }
 itertools       = { version = "0.12.0",  default-features = false, features = ["use_std"] }
 log             = { version = "0.4.17",  default-features = false }
-openssl         = { version = "0.10.59", default-features = false, features = ["vendored"]}
+openssl         = { version = "0.10.60", default-features = false, features = ["vendored"]}
 petgraph        = { version = "0.6.3",   default-features = false, features = ["serde-1"] }
 raqote          = { version = "0.8.2",   default-features = false, features = ["png"]}
 rand            = { version = "0.8.5",   default-features = false }
@@ -27,7 +27,7 @@ semver          = { version = "1.0.19",  default-features = false, features = ["
 serde           = { version = "1.0.163", default-features = false, features = ["derive"] }
 serde_json      = { version = "1.0.96",  default-features = false }
 strum           = { version = "0.25",    default-features = false, features = ["derive"] }
-tempfile        = { version = "3.5.0",   default-features = false }
+tempfile        = { version = "3.8.1",   default-features = false }
 tokio           = { version = "1.28.1",  default-features = false, features = ["macros", "rt-multi-thread"] }
 zstd            = { version = "0.13.0",  default-features = false }
