--- dataset/toy1/phylogeny.json
@@ -1,50 +1,0 @@
-{
-  "graph": {
-    "nodes": [
-      "root",
-      "A",
-      "B",
-      "C"
-    ],
-    "node_holes": [],
-    "edge_property": "directed",
-    "edges": [
-      [
-        0,
-        1,
-        1
-      ],
-      [
-        0,
-        2,
-        1
-      ],
-      [
-        1,
-        3,
-        1
-      ],
-      [
-        2,
-        3,
-        1
-      ]
-    ]
-  },
-  "order": [
-    "A",
-    "B",
-    "C"
-  ],
-  "recombinants": [
-    "C"
-  ],
-  "recombinants_all": [
-    "C"
-  ],
-  "non_recombinants_all": [
-    "root",
-    "A",
-    "B"
-  ]
-}
