--- dataset/toy1/populations.fasta
@@ -1,6 +1,0 @@
->A
-CCCCCCCCCC
->B
-TTTTTTTTTT
->C
-CCCCCCTTTT
