--- dataset/toy1/reference.fasta
@@ -1,2 +1,0 @@
->Reference
-AAAAAAAAAA
