+++ src/dataset/sarscov2/edge_cases.rs
@@ -1,28 +1,10 @@
 use crate::cli::run;
-//use crate::phylogeny::Phylogeny;
 use color_eyre::eyre::{Report, Result};
-//use itertools::Itertools;
 use log::debug;
 
 /// Create default SARS-CoV-2 recombinant edge cases.
 pub fn default() -> Result<Vec<run::Args>, Report> {
     let mut edge_cases: Vec<run::Args> = Vec::new();
-
-    // // --------------------------------------------------------------------
-    // // Designated Recombinants
-
-    // let manual = vec!["XCF", "XCG"].into_iter().map(String::from).collect_vec();
-
-    // for recombinant in &phylogeny.recombinants {
-    //     if manual.contains(recombinant) {
-    //         continue
-    //     }
-    //     debug!("Creating auto edge case: {recombinant}");
-    //     let mut edge_case = run::Args::default();
-    //     edge_case.population = Some(recombinant.to_string());
-    //     edge_case.parents = Some(phylogeny.get_parents(recombinant)?);
-    //     edge_cases.push(edge_case);
-    // }
 
     // --------------------------------------------------------------------
     // Manual
